graphql-backend-cats/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
strawberry-graphql[fastapi]==0.246.0
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
aiosqlite==0.19.0
requests==2.31.0
python-dotenv==1.0.0